# server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Photo Diary API",
    "description": "Self-hosted photo-diary backend. Schema is generated from TypeBox-validated route handlers.",
    "version": "0.16.0"
  },
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "JWT issued by POST /api/v1/tokens. Send as `Authorization: Bearer <token>`."
      }
    },
...
  },
  "paths": {
...
    "/api/v1/photos/query": {
      "post": {
        "summary": "Filter-aware cross-gallery photo list (admin)",
        "tags": [
          "photos"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "filter": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "array",
                        "items": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "number"
                            },
                            {
                              "type": "boolean"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        }
                      }
                    }
                  },
                  "dateRange": {
                    "type": "object",
                    "properties": {
                      "from": {
                        "type": "string",
                        "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                      },
                      "to": {
                        "type": "string",
                        "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                      }
                    },
                    "additionalProperties": false
                  },
                  "year": {
                    "type": "integer",
                    "minimum": 1900,
                    "maximum": 9999
                  },
                  "month": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 12
                  },
                  "day": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 31
                  },
                  "lang": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 8
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {},
                    "additionalProperties": true
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/photos/{photoId}": {
      "get": {
        "summary": "Get one photo by id",
        "tags": [
          "photos"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "photoId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "additionalProperties": true
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update one photo (admin)",
        "tags": [
          "photos"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "titleLocalized": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "string"
                    }
                  },
                  "descriptionLocalized": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "string"
                    }
                  },
                  "taken": {
                    "type": "object",
                    "properties": {
                      "author": {
                        "type": "string"
                      },
                      "location": {
                        "type": "object",
                        "properties": {
                          "country": {
                            "type": "string"
                          },
                          "place": {
                            "type": "string"
                          },
                          "placeLocalized": {
                            "type": "object",
                            "additionalProperties": {
                              "type": "string"
                            }
                          },
                          "coordinates": {
                            "type": "object",
                            "properties": {
                              "latitude": {
                                "anyOf": [
                                  {
                                    "type": "number"
                                  },
                                  {
                                    "type": "null"
                                  }
                                ]
                              },
                              "longitude": {
                                "anyOf": [
                                  {
                                    "type": "number"
                                  },
                                  {
                                    "type": "null"
                                  }
                                ]
                              },
                              "altitude": {
                                "anyOf": [
                                  {
                                    "type": "number"
                                  },
                                  {
                                    "type": "null"
                                  }
                                ]
                              }
                            },
                            "additionalProperties": false
                          }
                        },
                        "additionalProperties": false
                      }
                    },
                    "additionalProperties": false
                  },
                  "camera": {
                    "type": "object",
                    "properties": {
                      "make": {
                        "type": "string"
                      },
                      "model": {
                        "type": "string"
                      }
                    },
                    "additionalProperties": false
                  },
                  "lens": {
                    "type": "object",
                    "properties": {
                      "make": {
                        "type": "string"
                      },
                      "model": {
                        "type": "string"
                      }
                    },
                    "additionalProperties": false
                  },
                  "exposure": {
                    "type": "object",
                    "properties": {
                      "focalLength": {
                        "type": "number"
                      },
                      "focalLength35mmEquiv": {
                        "type": "number"
                      },
                      "aperture": {
                        "type": "number"
                      }
                    },
                    "additionalProperties": false
                  },
                  "isPrivate": {
                    "type": "boolean"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "photoId",
            "required": true
          }
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "delete": {
        "summary": "Delete one photo (admin)",
        "tags": [
          "photos"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "photoId",
            "required": true
          }
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/api/v1/gallery-photos/{galleryId}": {
      "get": {
        "summary": "List photos in a gallery",
        "tags": [
          "gallery-photos"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "lang",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "galleryId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {},
                    "additionalProperties": true
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/gallery-photos/{galleryId}/query": {
      "post": {
        "summary": "Filtered + view-scoped photo fetch",
        "tags": [
          "gallery-photos"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "filter": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "array",
                        "items": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "number"
                            },
                            {
                              "type": "boolean"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        }
                      }
                    }
                  },
                  "dateRange": {
                    "type": "object",
                    "properties": {
                      "from": {
                        "type": "string",
                        "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                      },
                      "to": {
                        "type": "string",
                        "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                      }
                    },
                    "additionalProperties": false
                  },
                  "numericRanges": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "object",
                      "properties": {
                        "min": {
                          "type": "number"
                        },
                        "max": {
                          "type": "number"
                        }
                      },
                      "additionalProperties": false
                    }
                  },
                  "year": {
                    "type": "integer",
                    "minimum": 1900,
                    "maximum": 9999
                  },
                  "month": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 12
                  },
                  "day": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 31
                  },
                  "lang": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 8
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "galleryId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {},
                    "additionalProperties": true
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/gallery-photos/{galleryId}/by-original-filename/{originalFilename}": {
      "get": {
        "summary": "Look up a photo by its original camera filename",
        "tags": [
          "gallery-photos"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "lang",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "galleryId",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "originalFilename",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "additionalProperties": true
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}